refactor: 1. 멀티 모듈 구조로 전환 2. DB를 외부 모듈로 분할 3. board 배포 스크립트 및 docker-compose 추가 4. 통합 배포

+++ auth/src/main/java/com/harpsharp/auth/service/UserService.java
@@ -1,5 +1,6 @@
 package com.harpsharp.auth.service;
 
+import com.harpsharp.auth.dto.JoinDTO;
 import com.harpsharp.infra_rds.entity.User;
 import com.harpsharp.infra_rds.dto.UserDTO;
 import com.harpsharp.auth.exceptions.UserAlreadyExistsException;
@@ -10,13 +11,51 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class UserService {
-    private final JwtUtil jwtUtil;
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    private final RefreshTokenService refreshTokenService;
+
+
+    public void registerUser(JoinDTO joinDTO, String role) {
+        String username = joinDTO.getUsername();
+        String password = joinDTO.getPassword();
+        String email = joinDTO.getEmail();
+        //String profile_image = joinDTO.getProfile_image();
+
+        if (userRepository.existsByEmail(email)) {
+            throw UserAlreadyExistsException.builder()
+                    .code("USER_ALREADY_EXISTS")
+                    .message("User with email " + joinDTO.getEmail() + " already exists.")
+                    .build();
+        }
+
+        if(userRepository.existsByUsername(username)) {
+            throw UserAlreadyExistsException.builder()
+                    .code("USER_ALREADY_EXISTS")
+                    .message("User with username " + joinDTO.getUsername() + " already exists.")
+                    .build();
+        }
+
+        User user = User.builder()
+                .username(username)
+                .password(passwordEncoder.encode(password))
+                .email(email)
+                .role(role)
+                .social_type("harp")
+                .build();
+
+        userRepository.save(user);
+    }
+
+    public void clear(){
+        userRepository.deleteAll();
+    }
 
     public void updateUser(Long userId, UserDTO updatedDTO){
         String updatedUsername = updatedDTO.getUsername();
@@ -61,11 +100,17 @@ public class UserService {
 
         if(updatedUser == null) throw new IllegalArgumentException();
 
-        ;
+
         userRepository.save(existUser.updateUser(updatedUser));
     }
 
-    public void deleteById(Long userId){
+    public Optional<User> findById(Long userId){ return userRepository.findById(userId); }
+
+    public void deleteById(Long userId, String accessToken){
+        if(!userRepository.existsById(userId)){ throw new IllegalArgumentException("존재하지 않는 유저입니다."); }
+
         userRepository.deleteById(userId);
+        refreshTokenService.deleteByToken(accessToken);
+
     }
 }
